Refactor: cleanup unnecessary articles in mock data
RISDEV-3510
